refactor(rename): feature-config module — DreamConfig→FeatureConfig, decisions→learning key
- dream-config.ts → feature-config.ts: FeatureConfig {memory,learning,knowledge}
- coerceConfig reads learning, coalesces from legacy decisions (decisions wins)
- writeConfig mkdirs .devflow/ root (config.json at neutral root)
- Repoint callers: init.ts, decisions.ts, memory.ts, knowledge/toggle.ts
- tests/dream-config.test.ts → tests/learning-config.test.ts + coalesce cases
